--- homework2/CarFactory/Models/BodyType/Convertible.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.BodyType
-{
-    public class Convertible : IBodyType
-    {
-        public string Name { get; } = "Convertible";
-    }
-}
