--- homework2/CarFactory/Models/BodyType/HatchBack.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.BodyType
-{
-    public class HatchBack : IBodyType
-    {
-        public string Name { get; } = "HatchBack";
-    }
-}
