--- homework2/CarFactory/Models/BodyType/IBodyType.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.BodyType
-{
-    public interface IBodyType
-    {
-        public string Name { get; }
-    }
-}
