--- homework2/CarFactory/Models/BodyType/Sedan.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.BodyType
-{
-    public class Sedan : IBodyType
-    {
-        public string Name { get; } = "Sedan";
-    }
-}
