+++ homework2/CarFactory/Models/BodyTypes/Convertible.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.BodyTypes
+{
+    public class Convertible : IBodyType
+    {
+        public string Name { get; } = "Convertible";
+    }
+}
