+++ homework2/CarFactory/Models/BodyTypes/HatchBack.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.BodyTypes
+{
+    public class HatchBack : IBodyType
+    {
+        public string Name { get; } = "HatchBack";
+    }
+}
