+++ homework2/CarFactory/Models/BodyTypes/IBodyType.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.BodyTypes
+{
+    public interface IBodyType
+    {
+        public string Name { get; }
+    }
+}
