+++ homework2/CarFactory/Models/BodyTypes/Sedan.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.BodyTypes
+{
+    public class Sedan : IBodyType
+    {
+        public string Name { get; } = "Sedan";
+    }
+}
