--- homework2/CarFactory/Models/Brand/Honda.cs
@@ -1,9 +1,0 @@
-﻿using CarFactory.Models.CarModel;
-
-namespace CarFactory.Models.Brand
-{
-    public class Honda : IBrand
-    {
-        public string Name { get; } = "Honda";
-    }
-}
