--- homework2/CarFactory/Models/Brand/IBrand.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.CarModel
-{
-    public interface IBrand
-    {
-        public string Name { get; }
-    }
-}
