--- homework2/CarFactory/Models/Brand/Mercedez.cs
@@ -1,9 +1,0 @@
-﻿using CarFactory.Models.CarModel;
-
-namespace CarFactory.Models.Brand
-{
-    public class Mercedez : IBrand
-    {
-        public string Name { get; } = "Mercedez";
-    }
-}
