--- homework2/CarFactory/Models/Brand/Mitsubishi.cs
@@ -1,9 +1,0 @@
-﻿using CarFactory.Models.CarModel;
-
-namespace CarFactory.Models.Brand
-{
-    public class Mitsubishi : IBrand
-    {
-        public string Name { get; } = "Mitsubishi";
-    }
-}
