--- homework2/CarFactory/Models/Brand/Toyota.cs
@@ -1,9 +1,0 @@
-﻿using CarFactory.Models.CarModel;
-
-namespace CarFactory.Models.Brand
-{
-    public class Toyota : IBrand
-    {
-        public string Name { get; } = "Toyota";
-    }
-}
