--- homework2/CarFactory/Models/Brand/Volkswagen.cs
@@ -1,9 +1,0 @@
-﻿using CarFactory.Models.CarModel;
-
-namespace CarFactory.Models.Brand
-{
-    public class Volkswagen : IBrand
-    {
-        public string Name { get; } = "Volkswagen";
-    }
-}
