+++ homework2/CarFactory/Models/Brands/Honda.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Brands
+{
+    public class Honda : IBrand
+    {
+        public string Name { get; } = "Honda";
+    }
+}
