+++ homework2/CarFactory/Models/Brands/IBrand.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Brands
+{
+    public interface IBrand
+    {
+        public string Name { get; }
+    }
+}
