+++ homework2/CarFactory/Models/Brands/Mercedez.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Brands
+{
+    public class Mercedez : IBrand
+    {
+        public string Name { get; } = "Mercedez";
+    }
+}
