+++ homework2/CarFactory/Models/Brands/Mitsubishi.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Brands
+{
+    public class Mitsubishi : IBrand
+    {
+        public string Name { get; } = "Mitsubishi";
+    }
+}
