+++ homework2/CarFactory/Models/Brands/Toyota.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Brands
+{
+    public class Toyota : IBrand
+    {
+        public string Name { get; } = "Toyota";
+    }
+}
