+++ homework2/CarFactory/Models/Brands/Volkswagen.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Brands
+{
+    public class Volkswagen : IBrand
+    {
+        public string Name { get; } = "Volkswagen";
+    }
+}
