--- homework2/CarFactory/Models/Color/Black.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.Color
-{
-    public class Black : IColor
-    {
-        public string Name { get; } = "Black";
-    }
-}
