--- homework2/CarFactory/Models/Color/Blue.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.Color
-{
-    public class Blue : IColor
-    {
-        public string Name { get; } = "Blue";
-    }
-}
