--- homework2/CarFactory/Models/Color/IColor.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.Color
-{
-    public interface IColor
-    {
-        public string Name { get; }
-    }
-}
