--- homework2/CarFactory/Models/Color/Red.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.Color
-{
-    public class Red : IColor
-    {
-        public string Name { get; } = "Red";
-    }
-}
