--- homework2/CarFactory/Models/Color/White.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.Color
-{
-    public class White : IColor
-    {
-        public string Name { get; } = "White";
-    }
-}
