+++ homework2/CarFactory/Models/Colors/Black.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Colors
+{
+    public class Black : IColor
+    {
+        public string Name { get; } = "Black";
+    }
+}
