+++ homework2/CarFactory/Models/Colors/Blue.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Colors
+{
+    public class Blue : IColor
+    {
+        public string Name { get; } = "Blue";
+    }
+}
