+++ homework2/CarFactory/Models/Colors/IColor.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Colors
+{
+    public interface IColor
+    {
+        public string Name { get; }
+    }
+}
