+++ homework2/CarFactory/Models/Colors/Red.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Colors
+{
+    public class Red : IColor
+    {
+        public string Name { get; } = "Red";
+    }
+}
