+++ homework2/CarFactory/Models/Colors/White.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Colors
+{
+    public class White : IColor
+    {
+        public string Name { get; } = "White";
+    }
+}
