--- homework2/CarFactory/Models/Engine/IEngine.cs
@@ -1,10 +1,0 @@
-﻿namespace CarFactory.Models.Engine
-{
-    public interface IEngine
-    {
-        public string Name { get; }
-        public int MaxSpeed { get; }
-        public int MaxGears { get; }
-        
-    }
-}
