--- homework2/CarFactory/Models/Engine/V12.cs
@@ -1,9 +1,0 @@
-﻿namespace CarFactory.Models.Engine
-{
-    public class V12 : IEngine
-    {
-        public string Name { get; } = "V12";
-        public int MaxSpeed { get; } = 217;
-        public int MaxGears { get; } = 7;
-    }
-}
