--- homework2/CarFactory/Models/Engine/V6.cs
@@ -1,9 +1,0 @@
-﻿namespace CarFactory.Models.Engine
-{
-    public class V6 : IEngine
-    {
-        public string Name { get; } = "V6";
-        public int MaxSpeed { get; } = 196;
-        public int MaxGears { get; } = 6;
-    }
-}
