--- homework2/CarFactory/Models/Engine/V8.cs
@@ -1,9 +1,0 @@
-﻿namespace CarFactory.Models.Engine
-{
-    public class V8 : IEngine
-    {
-        public string Name { get; } = "V8";
-        public int MaxSpeed { get; } = 211;
-        public int MaxGears { get; } = 7;
-    }
-}
