+++ homework2/CarFactory/Models/Engines/IEngine.cs
@@ -1,0 +1,10 @@
+﻿namespace CarFactory.Models.Engines
+{
+    public interface IEngine
+    {
+        public string Name { get; }
+        public int MaxSpeed { get; }
+        public int MaxGears { get; }
+        
+    }
+}
