+++ homework2/CarFactory/Models/Engines/V12.cs
@@ -1,0 +1,9 @@
+﻿namespace CarFactory.Models.Engines
+{
+    public class V12 : IEngine
+    {
+        public string Name { get; } = "V12";
+        public int MaxSpeed { get; } = 217;
+        public int MaxGears { get; } = 7;
+    }
+}
