--- homework2/CarFactory/Models/SteeringPosition/ISteeringPosition.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.SteeringPosition
-{
-    public interface ISteeringPosition
-    {
-        public string Name { get; }
-    }
-}
