--- homework2/CarFactory/Models/SteeringPosition/LeftPosition.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.SteeringPosition
-{
-    public class LeftPosition : ISteeringPosition
-    {
-        public string Name { get; } = "LeftPosition";
-    }
-}
