--- homework2/CarFactory/Models/SteeringPosition/RightPosition.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.SteeringPosition
-{
-    public class RightPosition : ISteeringPosition
-    {
-        public string Name { get; } = "RightPosition";
-    }
-}
