+++ homework2/CarFactory/Models/SteeringPositions/ISteeringPosition.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.SteeringPositions
+{
+    public interface ISteeringPosition
+    {
+        public string Name { get; }
+    }
+}
