+++ homework2/CarFactory/Models/SteeringPositions/LeftPosition.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.SteeringPositions
+{
+    public class LeftPosition : ISteeringPosition
+    {
+        public string Name { get; } = "LeftPosition";
+    }
+}
