+++ homework2/CarFactory/Models/SteeringPositions/RightPosition.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.SteeringPositions
+{
+    public class RightPosition : ISteeringPosition
+    {
+        public string Name { get; } = "RightPosition";
+    }
+}
