--- homework2/CarFactory/Models/Transmission/Automatical.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.Transmission
-{
-    public class Automatical : ITransmission
-    {
-        public string Name { get; } = "Automatical";
-    }
-}
