--- homework2/CarFactory/Models/Transmission/ITransmission.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.Transmission
-{
-    public interface ITransmission
-    {
-        public string Name { get; }
-    }
-}
