--- homework2/CarFactory/Models/Transmission/Mechanical.cs
@@ -1,7 +1,0 @@
-﻿namespace CarFactory.Models.Transmission
-{
-    public class Mechanical : ITransmission
-    {
-        public string Name { get; } = "Mechanical";
-    }
-}
