+++ homework2/CarFactory/Models/Transmissions/Automatical.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Transmissions
+{
+    public class Automatical : ITransmission
+    {
+        public string Name { get; } = "Automatical";
+    }
+}
