+++ homework2/CarFactory/Models/Transmissions/ITransmission.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Transmissions
+{
+    public interface ITransmission
+    {
+        public string Name { get; }
+    }
+}
