+++ homework2/CarFactory/Models/Transmissions/Mechanical.cs
@@ -1,0 +1,7 @@
+﻿namespace CarFactory.Models.Transmissions
+{
+    public class Mechanical : ITransmission
+    {
+        public string Name { get; } = "Mechanical";
+    }
+}
